+++ app/src/main/java/de/pawcode/cardstore/ui/components/LabelsListComponent.kt
@@ -8,10 +8,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.EditNote
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.FilledIconButton
 import androidx.compose.material3.FilterChip
@@ -22,6 +18,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -71,7 +68,7 @@ fun LabelsListComponent(
             },
             leadingIcon = {
               AnimatedVisibility(visible = chipSelected) {
-                Icon(Icons.Filled.Check, contentDescription = null)
+                Icon(painterResource(R.drawable.check_solid), contentDescription = null)
               }
             },
           )
@@ -93,7 +90,10 @@ fun LabelsListComponent(
       onClick = { onEdit() },
     ) {
       Icon(
-        imageVector = if (labels.isNotEmpty()) Icons.Filled.EditNote else Icons.Filled.Add,
+        painter =
+          painterResource(
+            if (labels.isNotEmpty()) R.drawable.edit_note_solid else R.drawable.add_solid
+          ),
         contentDescription = stringResource(R.string.labels_edit),
         modifier = Modifier.size(IconButtonDefaults.smallIconSize),
       )