+++ app/src/main/java/de/pawcode/cardstore/ui/components/SelectDropdownMenu.kt
@@ -1,16 +1,10 @@
 package de.pawcode.cardstore.ui.components
 
+import androidx.annotation.DrawableRes
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.Sort
-import androidx.compose.material.icons.automirrored.filled.TrendingUp
-import androidx.compose.material.icons.filled.AutoFixHigh
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.History
-import androidx.compose.material.icons.filled.SortByAlpha
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
@@ -25,7 +19,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
@@ -36,14 +30,14 @@ import de.pawcode.cardstore.data.enums.SortAttribute
 
 data class DropdownOption<TValue : Enum<TValue>>(
   val title: String,
-  val icon: ImageVector,
+  @param:DrawableRes val icon: Int,
   val value: TValue,
 )
 
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
 fun <TValue : Enum<TValue>> SelectDropdownMenu(
-  icon: ImageVector,
+  @DrawableRes icon: Int,
   title: String,
   value: TValue?,
   values: List<DropdownOption<TValue>>,
@@ -66,7 +60,7 @@ fun <TValue : Enum<TValue>> SelectDropdownMenu(
         ),
     ) {
       Icon(
-        icon,
+        painterResource(icon),
         contentDescription = title,
         modifier = Modifier.size(IconButtonDefaults.smallIconSize),
       )
@@ -88,10 +82,12 @@ fun <TValue : Enum<TValue>> SelectDropdownMenu(
               overflow = TextOverflow.Ellipsis,
             )
           },
-          leadingIcon = { Icon(imageVector = option.icon, contentDescription = option.title) },
+          leadingIcon = {
+            Icon(painter = painterResource(option.icon), contentDescription = option.title)
+          },
           trailingIcon = {
             AnimatedVisibility(visible = value === option.value) {
-              Icon(Icons.Filled.Check, contentDescription = null)
+              Icon(painterResource(R.drawable.check_solid), contentDescription = null)
             }
           },
           onClick = { onValueChange(option.value) },
@@ -105,29 +101,29 @@ fun <TValue : Enum<TValue>> SelectDropdownMenu(
 @Composable
 fun PreviewSelectDropdownMenu() {
   SelectDropdownMenu(
-    icon = Icons.AutoMirrored.Filled.Sort,
+    icon = R.drawable.sort_solid,
     title = stringResource(R.string.cards_sort),
     value = SortAttribute.INTELLIGENT,
     values =
       listOf(
         DropdownOption(
           title = stringResource(R.string.sort_intelligent),
-          icon = Icons.Filled.AutoFixHigh,
+          icon = R.drawable.wand_shine_solid,
           value = SortAttribute.INTELLIGENT,
         ),
         DropdownOption(
           title = stringResource(R.string.sort_alphabetically),
-          icon = Icons.Filled.SortByAlpha,
+          icon = R.drawable.sort_by_alpha_solid,
           value = SortAttribute.ALPHABETICALLY,
         ),
         DropdownOption(
           title = stringResource(R.string.sort_most_used),
-          icon = Icons.AutoMirrored.Filled.TrendingUp,
+          icon = R.drawable.trending_up_solid,
           value = SortAttribute.MOST_USED,
         ),
         DropdownOption(
           title = stringResource(R.string.sort_recently_used),
-          icon = Icons.Filled.History,
+          icon = R.drawable.history_solid,
           value = SortAttribute.RECENTLY_USED,
         ),
       ),