+++ app/src/main/java/de/pawcode/cardstore/ui/components/SettingsGroup.kt
@@ -4,15 +4,15 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Web
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import de.pawcode.cardstore.R
 
 @Composable
 fun SettingsGroup(modifier: Modifier = Modifier, title: String, content: @Composable () -> Unit) {
@@ -35,7 +35,7 @@ fun SettingsGroup(modifier: Modifier = Modifier, title: String, content: @Compos
 fun PreviewSettingsGroup() {
   SettingsGroup(title = "About") {
     SettingsItem(
-      icon = Icons.Default.Web,
+      icon = painterResource(R.drawable.web_solid),
       iconColor = MaterialTheme.colorScheme.onPrimaryFixedVariant,
       iconBackground = MaterialTheme.colorScheme.primaryFixed,
       title = "pawcode Development",
@@ -43,13 +43,13 @@ fun PreviewSettingsGroup() {
       onClick = {},
     )
     SettingsItem(
-      icon = Icons.Default.Web,
+      icon = painterResource(R.drawable.web_solid),
       iconColor = MaterialTheme.colorScheme.onSecondaryFixedVariant,
       iconBackground = MaterialTheme.colorScheme.secondaryFixed,
       title = "pawcode Development",
     )
     SettingsItem(
-      icon = Icons.Default.Web,
+      icon = painterResource(R.drawable.web_solid),
       iconColor = MaterialTheme.colorScheme.onTertiaryFixedVariant,
       iconBackground = MaterialTheme.colorScheme.tertiaryFixed,
       title = "pawcode Development",