+++ app/src/main/java/de/pawcode/cardstore/ui/components/SettingsItem.kt
@@ -7,29 +7,31 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Web
 import androidx.compose.material3.Icon
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import de.pawcode.cardstore.R
 
 @Composable
 fun SettingsItem(
-  icon: ImageVector,
+  icon: Painter,
   iconColor: Color,
   iconBackground: Color,
   title: String,
   subtitle: String? = null,
   onClick: (() -> Unit)? = null,
+  trailingContent: @Composable (() -> Unit)? = null,
 ) {
   ListItem(
     modifier =
@@ -38,6 +40,7 @@ fun SettingsItem(
     tonalElevation = 2.dp,
     headlineContent = { Text(text = title) },
     supportingContent = { subtitle?.let { Text(text = it) } },
+    trailingContent = trailingContent,
     leadingContent = {
       Box(
         modifier =
@@ -48,7 +51,7 @@ fun SettingsItem(
         contentAlignment = Alignment.Center,
       ) {
         Icon(
-          imageVector = icon,
+          painter = icon,
           contentDescription = null,
           tint = iconColor,
           modifier = Modifier.size(20.dp),
@@ -63,7 +66,7 @@ fun SettingsItem(
 fun PreviewSettingsItem() {
   Column {
     SettingsItem(
-      icon = Icons.Default.Web,
+      icon = painterResource(R.drawable.web_solid),
       iconColor = MaterialTheme.colorScheme.onPrimaryFixedVariant,
       iconBackground = MaterialTheme.colorScheme.primaryFixed,
       title = "pawcode Development",
@@ -71,17 +74,18 @@ fun PreviewSettingsItem() {
       onClick = {},
     )
     SettingsItem(
-      icon = Icons.Default.Web,
+      icon = painterResource(R.drawable.web_solid),
       iconColor = MaterialTheme.colorScheme.onSecondaryFixedVariant,
       iconBackground = MaterialTheme.colorScheme.secondaryFixed,
       title = "pawcode Development",
     )
     SettingsItem(
-      icon = Icons.Default.Web,
+      icon = painterResource(R.drawable.lock_solid),
       iconColor = MaterialTheme.colorScheme.onTertiaryFixedVariant,
       iconBackground = MaterialTheme.colorScheme.tertiaryFixed,
-      title = "pawcode Development",
-      subtitle = "View website",
+      title = "Restrict app access",
+      subtitle = "Lock app with biometric security",
+      trailingContent = { Switch(checked = false, onCheckedChange = {}) },
     )
   }
 }